pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policyforge"
version = "0.1.0"
description = "Local policy engine for AI agent tool-call gating with optional multi-cloud sync"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
aws = ["boto3>=1.28"]
azure = ["azure-identity>=1.17", "azure-storage-blob>=12.19"]
oci = ["oci>=2.112"]
all-clouds = ["boto3>=1.28", "azure-identity>=1.17", "azure-storage-blob>=12.19", "oci>=2.112"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pyyaml>=6.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "ruff>=0.4",
    "black>=24.0",
]

[tool.setuptools.packages.find]
include = ["policyforge*"]

[tool.setuptools.package-data]
policyforge = ["policies/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["boto3.*", "botocore.*", "oci.*", "azure.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.black]
target-version = ["py310"]
line-length = 99
